dynarr: ensure alignment takes header into account
When allocating a dynarr, the alignment must fit the header and not just
the array elements. This fix ensures the alignment is at least as wide
as the header.

Casting from a pointer to a dynarr header emits an alignment warning. As
long as we have aligned memory allocation (see the fix above), we should
have a properly aligned cast. The compiler does not know this so we
ignore the warning ("trust me bro") with an intermediate `void *` cast.